Don't paginate smaller datasets.
For smaller datasets, we would like to show the entire
dataset on the register page, rather than paginated
data. This lets us offer nice client-side sorting
and filtering on the dataset, giving the user a better
sense of the data.

Check the length of the dataset, and set pages to 0
if the dataset is under 10k rows. This is a configurable
option.